project-stedi/requirements.txt
awsglue-local==1.0.2
pyspark
